Added SwingyMonkeyLite; another 1K trials to the master Q-matrix
SwingyMonkeyLite works exactly like SwingyMonkey except that you
can't set a tick, and it doesn't draw any graphics to the screen,
so it goes much much faster. For compatibility, it does spawn a
small window with epoch no., so it still does require X,
unfortunately.

Another 1K trials -- current version is pvyh_1430779831.43 --
brings us up to an average of 88, with nowhere to go but up...